Ensure README.md is _NOT_ a symlink but resources/README.md _IS_
This is because github handles symlinks by rendering thier content,
which is a git text string pointing to the real path.

This causes problems under git mobile web client, as it shows only
the leading text of the README, and has a "view all of README.md" link
which then takes the user to the symlink content, not its destination.

Meaning, users can't read documentation through symlinks!